requirements.txt
azure-core==1.34.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
Flask==3.1.1
gunicorn==23.0.0
